Apply logic based on user_type
