+++ backend/app/mailers/user_mailer.rb
@@ -4,12 +4,12 @@ class UserMailer < ApplicationMailer
   def send_to_providers(subject, content)
     @content = content
 
-    mail(to: Subscriber.where(unsubscribed_at: nil).pluck(:email), subject: subject)
+    mail(to: Subscriber.where(user_type: "provider").where(unsubscribed_at: nil).pluck(:email), subject: subject)
   end
 
   def send_to_students(subject, content)
     @content = content
 
-    mail(to: Subscriber.where(unsubscribed_at: nil).pluck(:email), subject: subject)
+    mail(to: Subscriber.where(user_type: "student").where(unsubscribed_at: nil).pluck(:email), subject: subject)
   end
 end
