+++ backend/app/services/newsletter_service.rb
@@ -5,13 +5,15 @@ class NewsletterService
     @subject = params[:subject]
     @content = params[:content]
     @user_type = params[:user_type]
-    @subscribers_count = 0
   end
 
   def send_newsletter
-    if user_type == 'provider'
+    @subscribers = Subscriber.where(unsubscribed_at: nil)
+    has_no_user_type = @subscribers.where.not(user_type: user_type).count.zero?
+
+    if user_type == 'provider' && !has_no_user_type
       send_to_providers
-    elsif user_type == 'student'
+    elsif user_type == 'student' && !has_no_user_type
       send_to_students
     else
       return false
@@ -31,27 +33,27 @@ class NewsletterService
       user_type: @user_type
     )
 
-    subscribers = Subscriber.where(unsubscribed_at: nil)
-
-    subscribers.each do |subscriber|
+    @subscribers.where(user_type: user_type).each do |subscriber|
       NewsletterLog.create(
         newsletter: newsletter,
         email: subscriber.email
       )
     end
+
+    @id = newsletter.id
+
+    newsletter
   end
 
   def send_to_providers
     UserMailer.with(subject: subject, content: content)
-              .send_to_providers(@subject, @content).deliver_now
-
-    @subscribers_count = ActionMailer::Base.deliveries.size
+              .send_to_providers(@subject, @content)
+              .deliver_now
   end
 
   def send_to_students
     UserMailer.with(subject: subject, content: content)
-              .send_to_students.deliver_now
-
-    @subscribers_count = ActionMailer::Base.deliveries.size
+              .send_to_students(@subject, @content)
+              .deliver_now
   end
 end
